Remove initialization from Assistant, move to Python run script. Pythonic logging.
