+++ core/util/config.py
@@ -1,10 +1,14 @@
 import os
 import json
 
+import logging
+logger = logging.getLogger(__name__)
+
 class Config:
     """OPEN ASSISTANT CONFIGURATION"""
 
-    def __init__(self, path=None):
+    def __init__(self, path=None, **opts):
+        logger.info("Configuration")
         
         # DIRECTORIES
         self.cache_dir = os.path.join(path, 'cache')
@@ -29,11 +33,14 @@ class Config:
         self._make_dir(self.data_dir)
         
         self.options = self._read_options_file()
-        self.commands = self.options['commands']
+        self.commands = None
+        #self.create_strings_file()
+
 
     def _make_dir(self, directory):
         if not os.path.exists(directory):
             os.makedirs(directory)
+
 
     def _read_options_file(self):
         try:
@@ -42,7 +49,9 @@ class Config:
                 return _options
         except FileNotFoundError:
             # MAKE AN EMPTY OPTIONS NAMESPACE
+            print("No options file found: %s".format(self.opt_file))
             return {}
+
 
     def create_strings_file(self):
         # Open Strings File
@@ -54,3 +63,21 @@ class Config:
             for word in self.number_parser.number_words:
                 strings.write(word + " ")
             strings.write("\n")
+
+
+    #def update_voice_commands_if_changed(self):
+    #   """USE HASHES TO TEST IF THE VOICE COMMANDS HAVE CHANGED"""
+       #stored_hash = self.hasher['voice_commands']
+
+       # Calculate The Hash The Voice Commands Have Right Now
+       #hasher = self.hasher.get_hash_object()
+       #for voice_cmd in self.commands.keys():
+        #   hasher.update(voice_cmd.encode('utf-8'))
+           # Add A Separator To Avoid Odd Behavior
+        #   hasher.update('\n'.encode('utf-8'))
+       #new_hash = hasher.hexdigest()
+
+       #if new_hash != stored_hash:
+    #   self.create_strings_file()
+           #self.hasher['voice_commands'] = new_hash
+           #self.hasher.store()
